Add extra Admin stylesheet overrides to negate dark mode
Dark mode is nice, but it was clashing with some of the UI widgets,
making the RNA admin unusable and requiring users to flip to light mode
to do their work.

This change means we always show light-mode colours, even when dark mode is
active. The solution IS  heavy-handed, because it basically clones the
light-mode CSS and tells the browser to use that for dark mode, but it
works for now. Note that Django 4.0 seems to have a cleaner way to handle
dark-mode tweaks.

Resolves #631